# Feature/infer options arg