--- a/tests/test_generic_blendsql.py
+++ b/tests/test_generic_blendsql.py
@@ -1,22 +1,18 @@
 import pytest
 import pandas as pd
-import sqlite3
-from pathlib import Path
 from blendsql import blend
-from blendsql.db import SQLite
+from blendsql.db import Pandas
 from blendsql._exceptions import IngredientException, InvalidBlendSQL
+from tests.utils import select_first_option
 
 
 @pytest.fixture(scope="session")
-def db() -> SQLite:
-    """Create a dummy sqlite db to use in tests."""
-    dbpath = "./test_generic.db"
-    df = pd.DataFrame({"Name": ["Danny", "Emma", "Tony"], "Age": [23, 26, 19]})
-    con = sqlite3.connect(dbpath)
-    df.to_sql("w", con=con)
-    con.close()
-    yield SQLite(dbpath)
-    Path(dbpath).unlink()
+def db() -> Pandas:
+    """Create a dummy db to use in tests."""
+    return Pandas(
+        pd.DataFrame({"Name": ["Danny", "Emma", "Tony"], "Age": [23, 26, 19]}),
+        tablename="w",
+    )
 
 
 def test_error_on_delete1(db):
@@ -45,11 +41,30 @@ def test_error_on_delete2(db):
 
 def test_error_on_invalid_ingredient(db):
     blendsql = """
-    SELECT * transactions WHERE {{ingredient()}} = TRUE
+    SELECT * w WHERE {{ingredient()}} = TRUE
     """
     with pytest.raises(IngredientException):
         _ = blend(
             query=blendsql,
             db=db,
             ingredients={"This is not an ingredient type"},
+        )
+
+
+def test_error_on_bad_options_subquery(db):
+    blendsql = """
+    SELECT * FROM w 
+    WHERE {{
+        select_first_option(
+            'I am at a nice cafe right now',
+            'w::Name',
+            options=(SELECT * FROM w)
+        )
+    }}
+    """
+    with pytest.raises(InvalidBlendSQL):
+        _ = blend(
+            query=blendsql,
+            db=db,
+            ingredients={select_first_option},
         )