This PR extends the `infer_gen_constraints` function to the `options` arg of a `QAIngredient`. 

This minimizes verbosity in the common pattern where we have a `{column} = {QAIngredient}` predicate, and enforces a choice amongst the `column` values (unless otherwise specified).

Now, when we run `blend` with `infer_gen_constraints=True`, the below queries are equivalent.

```sql
SELECT * FROM w
    WHERE name = {{
        LLMQA(
            'Choose a name'
        )
    }}
```

```sql
SELECT * FROM w
    WHERE name = {{
        LLMQA(
            'Choose a name',
            options='w::name'
        )
    }}
```

## Benchmark Runtimes

The benchmarks saw no noticeable change in performance with this PR. A little uptick in the 'rugby' task, will keep an eye on that in the future.

Before:
| Task           |   Average Runtime |   # Unique Queries |
|:---------------|------------------:|-------------------:|
| financials     |         0.0414736 |                  7 |
| rugby          |         0.255618  |                  4 |
| 1966_nba_draft |         0.124267  |                  2 |

After:
| Task           |   Average Runtime |   # Unique Queries |
|:---------------|------------------:|-------------------:|
| financials     |         0.0461312 |                  7 |
| rugby          |         0.317933  |                  4 |
| 1966_nba_draft |         0.110849  |                  2 |